Client is no longer disconnecting, movement is bugged.

+++ protocol/src/main/java/net/md_5/bungee/protocol/packet/rewriter/serverside/ServerUpdateFoodRewriter.java
@@ -11,7 +11,7 @@ public class ServerUpdateFoodRewriter extends PacketRewriter
     public void rewrite(ByteBuf in, ByteBuf out)
     {
         out.writeBytes( in.readBytes( 4 ) );
-        DefinedPacket.writeVarInt( in.readInt(), out );
+        DefinedPacket.writeVarInt( in.readShort(), out );
         out.writeBytes( in.readBytes( in.readableBytes() ) );
     }
 }
